Homework_8 CustomLinkedList
Yuriy, I created the branch, but it is heavy I think it is big size as all of my lessons and HomeWorks are in. I did not want to have all this attached. I only wished to attach the homework_8. Maybe you can help me with that as you are more knowledgeable in this, please? Thank you!